LT-12: ErrorPageWrapper
