+++ packages/client/src/pages/404Page/ui/404Page.module.scss
@@ -14,19 +14,4 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  h1 {
-    color: $dark-blue;
-    font-size: 8rem;
-    margin-bottom: 2rem;
-  }
-  h2 {
-    color: $dark-blue;
-    font-size: 2rem;
-  }
-}
-
-.link {
-  font-family: $default-font-family;
-  font-size: 1.25rem;
-  margin-top: 2rem;
 }
