+++ packages/client/src/shared/ui/ErrorPageWrapper/ErrorPageWrapper.module.scss
@@ -1,0 +1,16 @@
+.title {
+  color: $dark-blue;
+  font-size: 8rem;
+  margin-bottom: 2rem;
+}
+
+.description {
+  color: $dark-blue;
+  font-size: 2rem;
+}
+
+.link {
+  font-family: $default-font-family;
+  font-size: 1.25rem;
+  margin-top: 2rem;
+}
