+++ packages/client/src/shared/ui/ErrorPageWrapper/types.tsx
@@ -1,0 +1,4 @@
+export interface ErrorPageWrapperProps {
+  title: string
+  description: string
+}
